+++ docs/changelog.rst
@@ -3,6 +3,13 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
+Version 1.0.3 (2025-10-11)
+---------------------------
+
+**Added**
+
+- Optional proxy configuration for single and per-scheme setups via ``GPlayScraper`` and ``Config``
+
 Version 1.0.2 (2025-10-11)
 ---------------------------